refactor: 상수 모듈 변경 - https://github.com/next-step/js-todo-list-step2/pull/24#discussion_r515945209
