--- src/constants/index.js
@@ -1,7 +1,0 @@
-import { filter } from './filter.js';
-
-export const ALL  = 'all';
-export const ACTIVE  = 'active';
-export const COMPLETED  = 'completed';
-
-export { filter };
