+++ src/template/Todo/TodoContainer.js
@@ -2,7 +2,7 @@ import TodoInput from './TodoInput.js';
 import TodoList from './TodoList.js';
 import TodoCountContainer from './TodoCountContainer.js';
 import { useState } from '../../core/state.js';
-import * as CONST from '../../constants/index.js';
+import * as CONST from '../../constants/filter.js';
 import { createDOM } from '../../utils.js';
 
 const TodoContainer = () => {
